+++ assets/js/workflow-diagram/types.ts
@@ -42,6 +42,7 @@ export namespace Lightning {
     edge?: boolean;
     error_path?: boolean;
     errors: any;
+    js_expression_label?: string;
   }
 
   export type Workflow = {