+++ test/lightning/workflows/edge_test.exs
@@ -35,8 +35,8 @@ defmodule Lightning.Workflows.EdgeTest do
       refute changeset.valid?
 
       assert {:condition,
-              {"must be :always when source is a trigger",
-               [validation: :inclusion, enum: [:always]]}} in changeset.errors
+              {"must be :always or :js_expression when source is a trigger",
+               [validation: :inclusion, enum: [:always, :js_expression]]}} in changeset.errors
     end
 
     test "can't have both source_job_id and source_trigger_id" do
@@ -200,6 +200,154 @@ defmodule Lightning.Workflows.EdgeTest do
       assert changeset.data.enabled ||
                Map.get(changeset.changes, :enabled, true),
              "Edges with a source_trigger_id should always be enabled"
+    end
+
+    test "requires js_expression condition to have a label and js body" do
+      changeset =
+        Edge.changeset(
+          %Edge{
+            id: Ecto.UUID.generate(),
+            workflow_id: Ecto.UUID.generate(),
+            source_job_id: Ecto.UUID.generate(),
+            enabled: true
+          },
+          %{condition: :js_expression}
+        )
+
+      assert changeset.errors == [
+               js_expression_label: {"can't be blank", [validation: :required]},
+               js_expression_body: {"can't be blank", [validation: :required]}
+             ]
+    end
+
+    test "requires js_expression label and condition to have limited length" do
+      changeset =
+        Edge.changeset(
+          %Edge{
+            id: Ecto.UUID.generate(),
+            workflow_id: Ecto.UUID.generate(),
+            source_job_id: Ecto.UUID.generate(),
+            enabled: true
+          },
+          %{
+            condition: :js_expression,
+            js_expression_label: String.duplicate("a", 256),
+            js_expression_body: String.duplicate("a", 256)
+          }
+        )
+
+      assert changeset.errors == [
+               js_expression_body: {
+                 "should be at most %{count} character(s)",
+                 [
+                   {:count, 255},
+                   {:validation, :length},
+                   {:kind, :max},
+                   {:type, :string}
+                 ]
+               },
+               js_expression_label:
+                 {"should be at most %{count} character(s)",
+                  [count: 255, validation: :length, kind: :max, type: :string]}
+             ]
+    end
+
+    test "requires JS expression to have valid syntax" do
+      edge = %Edge{
+        id: Ecto.UUID.generate(),
+        workflow_id: Ecto.UUID.generate(),
+        source_job_id: Ecto.UUID.generate(),
+        enabled: true
+      }
+
+      js_attrs = %{
+        condition: :js_expression,
+        js_expression_label: "Some JS Expression"
+      }
+
+      changeset =
+        Edge.changeset(
+          edge,
+          Map.put(
+            js_attrs,
+            :js_expression_body,
+            "state.data.foo == 'bar';"
+          )
+        )
+
+      assert changeset.errors == [
+               js_expression_body: {"must not contain a statement", []}
+             ]
+
+      changeset =
+        Edge.changeset(
+          edge,
+          Map.put(
+            js_attrs,
+            :js_expression_body,
+            "{ state.data.foo == 'bar' }"
+          )
+        )
+
+      assert changeset.errors == [
+               js_expression_body: {"must not contain a statement", []}
+             ]
+
+      changeset =
+        Edge.changeset(
+          edge,
+          Map.put(
+            js_attrs,
+            :js_expression_body,
+            "state.data.foo == 'bar' || state.data.bar == 'foo'"
+          )
+        )
+
+      assert changeset.errors == []
+    end
+
+    test "requires JS expression to have neither import or require statements" do
+      edge = %Edge{
+        id: Ecto.UUID.generate(),
+        workflow_id: Ecto.UUID.generate(),
+        source_job_id: Ecto.UUID.generate(),
+        enabled: true
+      }
+
+      js_attrs = %{
+        condition: :js_expression,
+        js_expression_label: "Some JS Expression"
+      }
+
+      changeset =
+        Edge.changeset(
+          edge,
+          Map.put(
+            js_attrs,
+            :js_expression_body,
+            "{ var fs = require('fs'); }"
+          )
+        )
+
+      assert changeset.errors == [
+               js_expression_body:
+                 {"must not contain import or require statements", []}
+             ]
+
+      changeset =
+        Edge.changeset(
+          edge,
+          Map.put(
+            js_attrs,
+            :js_expression_body,
+            "{ var fs = import('fs'); }"
+          )
+        )
+
+      assert changeset.errors == [
+               js_expression_body:
+                 {"must not contain import or require statements", []}
+             ]
     end
   end
 end